requirements.txt
arrapi==1.4.14
cloudscraper==1.2.71
GitPython==3.1.45
lxml==6.0.2
num2words==0.5.14
pathvalidate==3.3.1
pillow==11.3.0
PlexAPI==4.17.1
psutil==7.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pywin32==311; sys_platform == 'win32'
requests==2.32.5
ruamel.yaml==0.18.15
schedule==1.2.2
setuptools==80.9.0
tenacity==9.1.2
tmdbapis==1.2.30
fastapi
uvicorn
pytest
httpx
PyJWT
python-jose[cryptography]
passlib[bcrypt]
python-multipart
aiofiles
PyYAML
bcrypt==4.1.3